requirements.txt
absl-py==0.2.0
appnope==0.1.0
asn1crypto==0.22.0
astor==0.6.2
backcall==0.1.0
bleach==1.5.0
certifi==2018.4.16
cffi==1.10.0
cryptography==1.8.1
cycler==0.10.0
decorator==4.3.0
gast==0.2.0
grpcio==1.11.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipython-genutils==0.2.0
jedi==0.12.0
Keras==2.1.5
Markdown==2.6.11
numpy==1.14.2
packaging==16.8
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==17.0.0
pyparsing==2.2.0
python-dateutil==2.5.2
python-graph-core==1.8.2
pytz==2016.3
PyYAML==3.11
requests==2.14.2
scipy==1.0.1
simplegeneric==0.8.1
six==1.10.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
